requirements.txt
# --- Core API ---
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# --- Database (SQL Server + ORM) ---
SQLAlchemy==2.0.35
pyodbc==5.1.0
pydantic==2.9.2
pydantic-settings==2.3.4

# --- Migrations ---
alembic==1.13.3

# --- HTTP & Autodesk API Integration ---
requests==2.32.3
httpx==0.27.0

# --- Authentication & Security ---
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.3

# --- Logging & Utilities ---
loguru==0.7.2
uuid==1.30

# --- Optional (para pruebas o herramientas futuras) ---
pytest==8.3.3
black==24.10.0
isort==5.13.2